+++ examples/angular/src/app/app.component.ts
@@ -1,22 +1,23 @@
-import { Component } from '@angular/core';
-import { TranslationFunctions } from '../i18n/i18n-types';
-import { I18nService } from '../i18n/i18n.service';
+import { Component } from '@angular/core'
+import { TranslationFunctions } from '../i18n/i18n-types'
+import { I18nService } from '../i18n/i18n.service'
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+	selector: 'app-root',
+	templateUrl: './app.component.html',
+	styleUrls: ['./app.component.scss'],
 })
 export class AppComponent {
-  title = 'angular';
+	title = 'angular'
 
-  constructor(private i18nService: I18nService) {
-    i18nService.initI18n('en').then(() => {
-      // eslint-disable-next-line no-console
-      console.log(this.LL.STARTUP())
-    })
-  }
+	constructor(private i18nService: I18nService) {
+		i18nService.initI18n('en').then(() => {
+			// eslint-disable-next-line no-console
+			console.log(this.LL.STARTUP())
+		})
+	}
 
-  get LL(): TranslationFunctions { return this.i18nService.LL }
-
+	get LL(): TranslationFunctions {
+		return this.i18nService.LL
+	}
 }